pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mpt"
version = "0.0.2a"
description = "A module for Modified Mercle-Patrisa Tries."
authors = [{name = "Nynra", email="nynradev@pm.me"}]
readme = "README.md"
license = {text = "BSD 3-Clause License"}
classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies = [
    "cytoolz", "eth-hash", "eth-typing", "eth-utils",
    "pycryptodome", "rlp", "toolz",
    'importlib-metadata; python_version>"3.8"',
]

[tool.setuptools.packages.find]
where = ["src"]
